Add profile script in measurements/profile.rb.
Calculates the default example stub ten times. The stub is
calculated ten times instead of one so as to give the profiler
enough time to gather data.

The script accepts a single argument which will be used as the name
for the profiling data and PDF; supply this is you want to create a
new profile for comparison with a previous run. No argument will
just use "profile".

The output PDF is saved to measurements/ and opened automatically.