Optimise calculation process.
One of the bottlenecks in the calculation was that we called #always_on?
`Merit::POINTS * Order#producers#length` times. By pre-splitting the
producers into those which are always on, and those which aren't, we can
process each independently and avoid the predicate method.

This relies on Order#producers returning always_on producers before
those which aren't.

Before: 3.083563 seconds.
After:  2.441711 seconds.