Privatise LoadCurve#values.
"values" is part of the internal implementation, and having to set an
explicit value for each point makes it harder to break out of the
calculation loop earlier.

Now, if you want a specific point on the curve use #get:

    load_curve.get(point) # => 42.0

If the point has no value set, you'll get a sensible 0.0 back. To set a
value for a point, use #set:

    load_curve.set(point, value)

Finally, LoadCurve is now Enumerable so you can iterate through each
point, with nils automatically converted to 0.0:

    load_curve = LoadCurve.new([1, 2, nil, nil, ...])
    load_curve.map { |v| v } # => [1, 2, 0.0, 0.0, ...]