Add QuantizingCalculator.
This is a calculator which skips points in order to compute the result
faster.

This works on the principle that days which are close to one another are
expected to have very similar load characteristics; it is unlikely that
wind turbines will be in high demand on January 1st, but low demand on
the following day.

This calculator computes only a subset of days in the year, and then
copies the results for those days into the following few days. For
example, with a resolution of 3, it will compute the results for January
1st, then use the same figures for January 2nd and 3rd. January 4th will
then be computed, with its values being used for the 5th and 6th.

The result is that we're able to skip a large number of extra
calculations with only a very small loss of accuracy. "Freak
occurrences" in the load curves (such as an unusually high load in one
particular hour, on one day of the year) may be lost or exaggerated
depending on whether they happen on a computed day, but the overall
long-term trends remain.

Conflicts:
	spec/merit/calculator_spec.rb
	spec/merit/participants/producer_spec.rb