Add AveragingCalculator.
Creates a new AveragingCalculator which computes the production load for
each of the transient producers in a way which is faster than the
ordinary Calculator.

The total amount of demand assigned is very accurate, however since
the averaging process "smooths out" demand/load during the +@chunk_size+
period the load is often assigned to transient producers differently
than the ordinary Calculator.

On the upside, it is 5-6 times faster than Calculator.