Prevent breaking from assigning storage too soon
Due to floating point errors, "produced" may be a tiny but non-zero
value. This fooled the calculator into thinking that there was excess
must-run energy which couldn't be stored, and exited the must-run
calculation loop too soon (while there was in fact remaining storage
capacity, and other must-runs with excess).

Ref quintel/etmodel#2217