Allow SimpleReserve to track hourly volume
Allows SimpleReserve to keep track of the volume in each hour so that it
may be converted to an array.

SimpleReserve now has feature parity with Reserve, but still disallows
random access to methods like `set`, `add`, and `take`. The tradeoff is
that it is roughly twice as fast and should be preferred when random
access to the Reserve isn't needed.

    Calculating -------------------------------------
    SimpleReserve without decay
                            615.528  (± 2.9%) i/s -      3.087k in   5.019511s
    SimpleReserve with decay
                            571.618  (± 3.0%) i/s -      2.907k in   5.090073s
    SimpleReserve without decay, every 4th hour
                              1.095k (± 4.3%) i/s -      5.546k in   5.075598s
    Reserve without decay
                            323.333  (± 3.1%) i/s -      1.624k in   5.027766s
      Reserve with decay
                            249.778  (± 3.6%) i/s -      1.265k in   5.070824s
    Reserve without decay, every 4th hour
                            353.168  (± 4.0%) i/s -      1.785k in   5.062288s