Clean up and optimise price-sensitive flex
* Pre-groups flexibility options by price when all technologies have
  fixed pricing.

* When technologies have variable pricing, groups have to be re-created
  in every hour. This is both slow and leads to a significant number of
  object allocations. Variable pricing should be avoided with flex
  technologies when at all possible.

* Flex::Group has been removed - since technologies are always sorted by
  their cost rather than with explicit groups. GroupSet and
  FlexListBuilder are also no longer required, and have been removed.

* A new Flex::Group class groups flex technologies which have the same
  cost in an hour.

For a demanding real-world scenario, this improves performance of the
calculation from 400-600ms to a consistent 280ms. This is still
substantially worse than the 150ms taken prior to the introduction of
price-sensitive flex.

The use of Util.slice_each seems to result in many object allocations,
for reasons that aren't yet clear. Replacing these calls with a while
loop and manually incremented index fixed the problem, at the cost of
ugly code.