Don't assign A/O to flex when their cost is zero
When a flex technology marginal cost (max consumption price) is zero, do
not assign it any energy from always-on production. This may be the case
when exporting energy from always-ons and the foreign market has a price
of 0.

In these situations, we don't want to export as it would not be
profitable to do so.

This affects all flex technologies as:

1. We'll add non-zero max_consumption_price attributes to all flex
   technologies.
2. It avoids the need for an explosion of special-cases in Group to
   determine when and how to assign energy to participants which have
   the same price.