Support optimizing storage in the price curve
OptimizingStorage in Merit is implemented as a seprate producer and
consumer, both of which are inflexible (mandatory) loads. However, as
they represent a priced technology they must be accounted for when
calculating a price curve.

This commit implements subclasses of CurveProducer and User::WithCurve
which may have prices. PriceCurve detects the presence of one or more of
these participants and falls back to using them as pricing when a
price-sensitive or dispatchable would not otherwise be price-setting.

These markers take the price of other price-sensitives or dispatchables
and over-cut or under-cut respectively the price of the other
participants.