Fix incorrect deficit in price curve
When the most expensive dispatchable is fully-loaded, there exists a
deficit. However, this fails to account for the fact that the most
expensive dispatchable may have no available capacity; this is the case
if its capacity is set with a cuve, for example export with an
availability curve.

In those cases, we must find the most expensive available dispatchable,
and compare its actual and max output instead.